HHH-9840 Refactor org.hibernate.cache.spi.CacheKey into an interface
