+++ hibernate-core/src/main/java/org/hibernate/cache/spi/CacheKey.java
@@ -8,10 +8,6 @@ package org.hibernate.cache.spi;
 
 import java.io.Serializable;
 
-import org.hibernate.engine.spi.SessionFactoryImplementor;
-import org.hibernate.internal.util.compare.EqualsHelper;
-import org.hibernate.type.Type;
-
 /**
  * Allows multiple entity classes / collection roles to be stored in the same cache region. Also allows for composite
  * keys which do not properly implement equals()/hashCode().
@@ -19,81 +15,12 @@ import org.hibernate.type.Type;
  * @author Gavin King
  * @author Steve Ebersole
  */
-public class CacheKey implements Serializable {
-	private final Serializable key;
-	private final Type type;
-	private final String entityOrRoleName;
-	private final String tenantId;
-	private final int hashCode;
+public interface CacheKey {
+	
+	public Serializable getKey();
 
-	/**
-	 * Construct a new key for a collection or entity instance.
-	 * Note that an entity name should always be the root entity
-	 * name, not a subclass entity name.
-	 *
-	 * @param id The identifier associated with the cached data
-	 * @param type The Hibernate type mapping
-	 * @param entityOrRoleName The entity or collection-role name.
-	 * @param tenantId The tenant identifier associated this data.
-	 * @param factory The session factory for which we are caching
-	 */
-	public CacheKey(
-			final Serializable id,
-			final Type type,
-			final String entityOrRoleName,
-			final String tenantId,
-			final SessionFactoryImplementor factory) {
-		this.key = id;
-		this.type = type;
-		this.entityOrRoleName = entityOrRoleName;
-		this.tenantId = tenantId;
-		this.hashCode = calculateHashCode( type, factory );
-	}
+	public String getEntityOrRoleName();
 
-	private int calculateHashCode(Type type, SessionFactoryImplementor factory) {
-		int result = type.getHashCode( key, factory );
-		result = 31 * result + (tenantId != null ? tenantId.hashCode() : 0);
-		return result;
-	}
+	public String getTenantId();
 
-	public Serializable getKey() {
-		return key;
-	}
-
-	public String getEntityOrRoleName() {
-		return entityOrRoleName;
-	}
-
-	public String getTenantId() {
-		return tenantId;
-	}
-
-	@Override
-	public boolean equals(Object other) {
-		if ( other == null ) {
-			return false;
-		}
-		if ( this == other ) {
-			return true;
-		}
-		if ( hashCode != other.hashCode() || !( other instanceof CacheKey ) ) {
-			//hashCode is part of this check since it is pre-calculated and hash must match for equals to be true
-			return false;
-		}
-		final CacheKey that = (CacheKey) other;
-		return EqualsHelper.equals( entityOrRoleName, that.entityOrRoleName )
-				&& type.isEqual( key, that.key )
-				&& EqualsHelper.equals( tenantId, that.tenantId );
-	}
-
-	@Override
-	public int hashCode() {
-		return hashCode;
-	}
-
-	@Override
-	public String toString() {
-		// Used to be required for OSCache
-		return entityOrRoleName + '#' + key.toString();
-	}
 }
